usage_analytics/create_analytics/requirements.txt
google-cloud-bigquery===3.17.2
functions-framework===3.5.0
